requirements.txt
numpy
scipy
pandas
matplotlib
openpyxl
pyyaml
pytest
